Release 2.1.4
Release notes are available at:
https://docs.dremio.com/release-notes/21-release-notes.html

+++ sabot/kernel/src/main/java/com/dremio/exec/physical/config/HashAggregate.java
@@ -67,7 +67,7 @@ public class HashAggregate extends AbstractSingle {
     return aggrExprs;
   }
 
-  public double getCardinality() {
+  public float getCardinality() {
     return cardinality;
   }
